+++ test/vm/assign.dart
@@ -5,7 +5,7 @@ import 'package:test/test.dart';
 void main() {
   group('BitManipulation', () {
     test('SetBit', () {
-      final five = new BigInteger()..assignInt = 5;
+      final five = new BigInt()..assignInt = 5;
       expect(five.value, 5);
     });
   });
